Move compare2 file changes to compare, add demonstration notebook
